+++ 04/Parse.h
@@ -24,7 +24,7 @@ int toInt(std::string_view view) {
       throw std::invalid_argument("can't convert std::string_view to int.");
     }
 
-    num = num * 10 + static_cast<int>(digit - '0');
+    num = num * 10 + (digit - '0');
   }
 
   return num;
